flask_app/requirements.txt
flask
azure-eventhub
azure-storage-blob
pandas
pyarrow
gunicorn
